feat: move most introduction tour logic into IntroductionTour
